Created event.js and updated Trip Model
Added event.js in the api folder and made a POST request route where we can add a event into a specific trip based on the tripID in the params.

Updated Trip model to export Housing, Restaurant, and Event schemas for their respective apis to access and to create off of.

Routed event api in server.js to be "api/event"